requirements.txt
Pillow
numpy
opencv-python
matplotlib
torch
torchvision
scikit-image